fix: fix authorization header on getUserSessionStatus and change to a server action
